# backend/docs/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Paywall IAP API
  version: 1.0.0
  description: OpenAPI schema assembled from the current Gin router, handler DTOs, and advanced bandit endpoints.
servers:
  - url: /
...
paths:
...
  /v1/subscription:
    get:
      tags: [subscription]
      summary: Get current subscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [subscription]
      summary: Cancel current subscription
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Subscription canceled
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/subscription/restore:
    post:
      tags: [subscription]
      summary: Restore purchases and reactivate subscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No previous purchases found to restore
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/subscription/access:
    get:
      tags: [subscription]
      summary: Check premium access
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Access check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Meta:
      type: object
      required: [request_id, timestamp]
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
...
    SubscriptionResponse:
      type: object
      required: [id, status, source, platform, product_id, plan_type, expires_at, auto_renew, created_at, updated_at]
      properties:
        id: { type: string }
        status: { type: string }
        source: { type: string }
        platform: { type: string }
        product_id: { type: string }
        plan_type: { type: string }
        expires_at: { type: string }
        auto_renew: { type: boolean }
        created_at: { type: string }
        updated_at: { type: string }
    AccessCheckResponse:
      type: object
      required: [has_access]
      properties:
        has_access: { type: boolean }
        expires_at: { type: string }
        reason: { type: string }
...
    SubscriptionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/SubscriptionResponse'
        meta:
          $ref: '#/components/schemas/Meta'
    AccessCheckEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/AccessCheckResponse'
        meta:
          $ref: '#/components/schemas/Meta'
...