# backend/internal/interfaces/http/handlers/subscription.go
package handlers

import (
	"errors"

	"github.com/gin-gonic/gin"
	"github.com/bivex/paywall-iap/internal/application/command"
	"github.com/bivex/paywall-iap/internal/application/middleware"
	"github.com/bivex/paywall-iap/internal/application/query"
	domainErrors "github.com/bivex/paywall-iap/internal/domain/errors"
	"github.com/bivex/paywall-iap/internal/interfaces/http/response"
)

// SubscriptionHandler handles subscription endpoints
type SubscriptionHandler struct {
	getSubQuery         *query.GetSubscriptionQuery
	checkAccessQuery    *query.CheckAccessQuery
	cancelCmd           *command.CancelSubscriptionCommand
	restoreCmd          *command.RestoreSubscriptionCommand
	jwtMiddleware       *middleware.JWTMiddleware
}

// NewSubscriptionHandler creates a new subscription handler
func NewSubscriptionHandler(
	getSubQuery *query.GetSubscriptionQuery,
	checkAccessQuery *query.CheckAccessQuery,
	cancelCmd *command.CancelSubscriptionCommand,
	restoreCmd *command.RestoreSubscriptionCommand,
	jwtMiddleware *middleware.JWTMiddleware,
) *SubscriptionHandler {
	return &SubscriptionHandler{
		getSubQuery:      getSubQuery,
		checkAccessQuery: checkAccessQuery,
		cancelCmd:        cancelCmd,
		restoreCmd:       restoreCmd,
		jwtMiddleware:    jwtMiddleware,
	}
}

// GetSubscription returns the user's subscription details
// @Summary Get subscription details
// @Tags subscription
// @Produce json
// @Security Bearer
// @Success 200 {object} response.SuccessResponse{data=dto.SubscriptionResponse}
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Router /subscription [get]
func (h *SubscriptionHandler) GetSubscription(c *gin.Context) {
	userID := c.GetString("user_id")
	if userID == "" {
		response.Unauthorized(c, "User not authenticated")
		return
	}

	resp, err := h.getSubQuery.Execute(c.Request.Context(), userID)
	if err != nil {
		response.NotFound(c, "Subscription not found")
		return
	}

	response.OK(c, resp)
}

// CheckAccess checks if user has access to premium content
// @Summary Check access to premium content
// @Tags subscription
// @Produce json
// @Security Bearer
// @Success 200 {object} response.SuccessResponse{data=dto.AccessCheckResponse}
// @Failure 401 {object} response.ErrorResponse
// @Router /subscription/access [get]
func (h *SubscriptionHandler) CheckAccess(c *gin.Context) {
	userID := c.GetString("user_id")
	if userID == "" {
		response.Unauthorized(c, "User not authenticated")
		return
	}

	resp, err := h.checkAccessQuery.Execute(c.Request.Context(), userID)
	if err != nil {
		response.InternalError(c, "Failed to check access")
		return
	}

	response.OK(c, resp)
}

// CancelSubscription cancels the user's subscription
// @Summary Cancel subscription
// @Tags subscription
// @Accept json
// @Produce json
// @Security Bearer
// @Success 204
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Router /subscription [delete]
func (h *SubscriptionHandler) CancelSubscription(c *gin.Context) {
	userID := c.GetString("user_id")
	if userID == "" {
		response.Unauthorized(c, "User not authenticated")
		return
	}

	if err := h.cancelCmd.Execute(c.Request.Context(), userID); err != nil {
		if errors.Is(err, domainErrors.ErrSubscriptionNotActive) || errors.Is(err, domainErrors.ErrSubscriptionNotFound) {
			response.NotFound(c, "No active subscription found")
			return
		}
		response.InternalError(c, "Failed to cancel subscription")
		return
	}

	response.NoContent(c)
}

// RestoreSubscription restores the user's previous purchases/subscriptions
// @Summary Restore purchases
// @Tags subscription
// @Produce json
// @Security Bearer
// @Success 200 {object} response.SuccessResponse{data=dto.SubscriptionResponse}
// @Failure 401 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Router /subscription/restore [post]
func (h *SubscriptionHandler) RestoreSubscription(c *gin.Context) {
	userID := c.GetString("user_id")
	if userID == "" {
		response.Unauthorized(c, "User not authenticated")
		return
	}

	if h.restoreCmd == nil {
		response.InternalError(c, "Restore command not configured")
		return
	}

	resp, err := h.restoreCmd.Execute(c.Request.Context(), userID)
	if err != nil {
		if errors.Is(err, domainErrors.ErrSubscriptionNotFound) {
			response.NotFound(c, "No previous purchases found to restore")
			return
		}
		response.InternalError(c, "Failed to restore purchases: "+err.Error())
		return
	}

	response.OK(c, resp)
}